backend/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
djangorestframework-simplejwt
django-cors-headers